+++ scripts/make_travisconf.py
@@ -9,6 +9,7 @@ cfg = {}
 
 cfg['sudo'] = True
 cfg['language'] = 'python'
+cfg['cache'] = 'pip'
 
 cfg['git'] = {
     'submodules': False
@@ -26,48 +27,62 @@ make -e install-dev;
 make -e install-$BUILD;
 """]
 
-cfg['script'] = ["make -e $BUILD"]
-cfg['after_script'] = ["make -e after-$BUILD"]
+script = """
+if [ "$TRAVIS_PULL_REQUEST" = "false" ] || [ "$BUILD_PRS" != "false" ];
+then {};
+fi""".format
+
+cfg['script'] = [script("make -e $BUILD")]
 
 matrix = []
 cfg['matrix'] = {'include': matrix}
 
 matrix.append({
     'python': latest_python,
-    'env': 'BUILD=style'
+    'env': 'BUILD=style BUILD_PRS=true'
 })
 
 
-for python in python_versions:
-    if python == latest_python:
-        dav_servers = ("skip", "radicale", "owncloud", "nextcloud", "baikal",
-                       "davical", "icloud")
-        rs_servers = ()
-    else:
-        dav_servers = ("skip", "radicale")
-        rs_servers = ()
+for python, requirements in itertools.product(python_versions,
+                                              ("devel", "release", "minimal")):
+    dav_servers = ("radicale",)
+    rs_servers = ()
+    if python == latest_python and requirements == "release":
+        dav_servers += ("owncloud", "nextcloud", "baikal", "davical", "icloud",
+                        "fastmail")
+    elif requirements == "devel":
+        dav_servers += ("xandikos",)
 
-    for (server_type, server), requirements in itertools.product(
-        itertools.chain(
-            (("REMOTESTORAGE", x) for x in rs_servers),
-            (("DAV", x) for x in dav_servers)
-        ),
-        ("devel", "release", "minimal")
+    for server_type, server in itertools.chain(
+        (("REMOTESTORAGE", x) for x in rs_servers),
+        (("DAV", x) for x in dav_servers)
     ):
+
+        build_prs = server not in ("fastmail", "davical", "icloud")
         matrix.append({
             'python': python,
             'env': ("BUILD=test "
                     "{server_type}_SERVER={server} "
                     "REQUIREMENTS={requirements} "
+                    "BUILD_PRS={build_prs} "
                     .format(server_type=server_type,
                             server=server,
-                            requirements=requirements))
+                            requirements=requirements,
+                            build_prs=build_prs and "true" or "false"))
         })
+
+matrix.append({
+    'python': latest_python,
+    'env': ("BUILD=test "
+            "ETESYNC_TESTS=true "
+            "REQUIREMENTS=latest "
+            "BUILD_PRS=true ")
+})
 
 matrix.append({
     'language': 'generic',
     'os': 'osx',
-    'env': 'BUILD=test'
+    'env': 'BUILD=test BUILD_PRS=true'
 })
 
 json.dump(cfg, sys.stdout, sort_keys=True, indent=2)