+++ tests/storage/__init__.py
@@ -203,8 +203,9 @@ class StorageTests(object):
         if getattr(self, 'dav_server', '') == 'radicale':
             pytest.skip('MKCOL is broken under Radicale 1.x')
 
-        if getattr(self, 'dav_server', '') == 'icloud':
-            pytest.skip('iCloud requires a minimum-length for collection name')
+        if getattr(self, 'dav_server', '') in \
+           ('icloud', 'fastmail', 'davical'):
+            pytest.skip('Manual cleanup would be necessary.')
 
         args = get_storage_args(collection=None)
         args['collection'] = 'test'
@@ -225,6 +226,9 @@ class StorageTests(object):
         assert 'collection argument must not be given' in str(excinfo.value)
 
     def test_collection_arg(self, get_storage_args):
+        if self.storage_class.storage_name.startswith('etesync'):
+            pytest.skip('etesync uses UUIDs.')
+
         if self.supports_collections:
             s = self.storage_class(**get_storage_args(collection='test2'))
             # Can't do stronger assertion because of radicale, which needs a
@@ -249,8 +253,8 @@ class StorageTests(object):
                           get_storage_args, get_item):
         if getattr(self, 'dav_server', '') == 'radicale':
             pytest.skip('Radicale is fundamentally broken.')
-        if getattr(self, 'dav_server', '') == 'icloud':
-            pytest.skip('iCloud rejects uploads.')
+        if getattr(self, 'dav_server', '') in ('icloud', 'fastmail'):
+            pytest.skip('iCloud and FastMail reject this name.')
 
         monkeypatch.setattr('vdirsyncer.utils.generate_href', lambda x: x)
 
@@ -268,6 +272,10 @@ class StorageTests(object):
 
         (_, etag3), = s.list()
         assert etag2 == etag3
+
+        # etesync uses UUIDs for collection names
+        if self.storage_class.storage_name.startswith('etesync'):
+            return
 
         assert collection in urlunquote(s.collection)
         if self.storage_class.storage_name.endswith('dav'):
@@ -309,7 +317,7 @@ class StorageTests(object):
         if item_type != 'VEVENT':
             pytest.skip('This storage instance doesn\'t support iCalendar.')
 
-        uid = u'abc123'
+        uid = str(uuid.uuid4())
         item = Item(textwrap.dedent(u'''
         BEGIN:VCALENDAR
         VERSION:2.0