+++ CHANGELOG.rst
@@ -9,8 +9,28 @@ Package maintainers and users who have to manually update their installation
 may want to subscribe to `GitHub's tag feed
 <https://github.com/pimutils/vdirsyncer/tags.atom>`_.
 
-Version 0.15
-============
+Version 0.16.0
+==============
+
+- Strip ``METHOD:PUBLISH`` added by some calendar providers, see :gh:`502`.
+- Fix crash of Google storages when saving token file.
+- Make DAV discovery more RFC-conformant, see :ghpr:`585`.
+- Vdirsyncer is now tested against Xandikos, see :ghpr:`601`.
+- Subfolders with a leading dot are now ignored during discover for
+  ``filesystem`` storage.
+- Statuses are now stored in a sqlite database. Old data is automatically
+  migrated. Users with really large datasets should encounter performance
+  improvements. This means that **sqlite3 is now a dependency of vdirsyncer**.
+- Vdirsyncer now ignores folders with a leading dot. This makes it easier to
+  combine it with version control.
+- **Vdirsyncer is now licensed under the 3-clause BSD license**, see :gh:`610`.
+- Vdirsyncer now includes experimental support for `EteSync
+  <https://www.etesync.com/>`_, see :ghpr:`614`.
+
+Version 0.15.0
+==============
+
+*released on 28 February 2017*
 
 - Deprecated syntax for configuration values is now completely rejected. All
   values now have to be valid JSON.
@@ -56,7 +76,7 @@ Version 0.13.0
 - Config sections are now checked for duplicate names. This also means that you
   cannot have a storage section ``[storage foo]`` and a pair ``[pair foo]`` in
   your config, they have to have different names. This is done such that
-  console output is always unambigous. See :gh:`459`.
+  console output is always unambiguous. See :gh:`459`.
 - Custom commands can now be used for conflict resolution during sync. See
   :gh:`127`.
 - :storage:`http` now completely ignores UIDs. This avoids a lot of unnecessary
@@ -126,7 +146,7 @@ Version 0.11.0
   - The ``keyring`` fetching strategy has been dropped some versions ago, but
     the specialized error message has been dropped.
 
-  - A old status format from version 0.4 is no longer supported. If you're
+  - An old status format from version 0.4 is no longer supported. If you're
     experiencing problems, just delete your status folder.
 
 Version 0.10.0
@@ -225,7 +245,7 @@ Version 0.7.4
 - Improved error messages instead of faulty server behavior, see :gh:`290` and
   :gh:`300`.
 - Safer shutdown of threadpool, avoid exceptions, see :gh:`291`.
-- Fix a sync bug for read-only storages see commmit
+- Fix a sync bug for read-only storages see commit
   ``ed22764921b2e5bf6a934cf14aa9c5fede804d8e``.
 - Etag changes are no longer sufficient to trigger sync operations. An actual
   content change is also necessary. See :gh:`257`.